# openapi/swagger.yaml
swagger: '2.0'
info:
  description: >-
    Create a browser request and a RESTful API that gets all tickets (IT
    tickets) for a region, organized in a way suitable for generating a heat
    map.
  version: 1.0.0
  title: Issues by region API
  contact:
    email: mark.grand@hcl.com
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
host: localhost
basePath: /v2
tags:
  - name: itTicket
    description: an IT ticket
schemes:
  - https
paths:
  /ticket/heatmap:
    get:
      tags:
        - itTicket
      summary: Get summary ticket information
      description: Retrieve tickets with just fields useful for a heat map.
      operationId: getTicketsForHeatmap
      produces:
        - application/json
      parameters:
        - name: status
          in: query
          description: Ticket statuses to include.
          required: false
          type: string
          enum:
            - any
            - closed
            - open
          default: any
        - name: regions
          in: header
          required: true
          description: identify the regions to be returned
          type: array
          items:
            type: string
          collectionFormat: csv
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Ticket'
        '400':
          description: Invalid status value
definitions:
  Ticket:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier for ticket
      status: 
        $ref: '#/definitions/Status'
      location:
        $ref: '#/definitions/Geocode'
      urgency:
        type: integer
        minimum: 1
        maximum: 100
    required: 
      - id
      - status
      - location
  Geocode:
    type: object
    description: Latitude and longitude
    properties:
      latitude:
        type: number
        format: double
        maximum: 90.0
        minimum: -90.0
      longitude:
        type: number
        format: double
        maximum: 180.0
        minimum: -180.0
  Status:
    type: string
    enum:
      - closed
      - open